[BE] feat: 화가, 장르, 시대별 그림 조회와 조회수, 좋아요수, 변환수 내림차순 조회 기능 구현

See merge request s07-bigdata-recom-sub2/S07P22C104!73

+++ backend/business-service/src/main/java/com/artrend/businessservice/domain/painting/dto/LikedPaintingDto.java
@@ -20,9 +20,9 @@ public class LikedPaintingDto {
     private String medium;
     private String place;
     private String url;
-    private Long hits;
-    private Long totalLikeCount;
-    private Long totalChangeCount;
+    private Integer hits;
+    private Integer totalLikeCount;
+    private Integer totalChangeCount;
 
     @QueryProjection
     public LikedPaintingDto(Painting painting) {
